src/backend/pyproject.toml
[project]
name = "german-tutor-backend"
version = "0.1.0"
description = "AI German Tutor Backend - FastAPI WebSocket relay to Azure OpenAI"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "websockets>=14.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "python-dotenv>=1.0.0",
    "azure-identity>=1.19.0",
    "azure-cosmos>=4.7.0",
    "openai>=1.58.0",
    "httpx>=0.28.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
